Implement a RAM512 chip
Pretty much same strategy as for RAM64 (and RAM8) chip: use 8 chips
(RAM64 chips in this case) as the core. Hook input directly
to those chips. Use a DMux8Way chip to determine which RAM chip
will be loaded. Use a Mux8Way16 to select the correct RAM chip to
output.