Implement a Program Counter (PC) chip
Building this chip in the reverse order of the rules above.
"reset" takes priority so it needs to be the final step before the
output, while the final rule needs to be the first chip used so
that it applies first and can be overridden by the other rules.
I'm using a true value for the register's load input which feels... weird since
it's not loading conditionally on anything but in this case I think
that's the correct behaviour.