Proj 7: Implement most arithmetic/logical commands
Still working on the VM translator for the week 7 project. Once I
figured out the main patterns used in the Hack assembly language, these
set of commands have been pretty straight-forward. I've been testing
some of them on the CPU emulator which is why there are some .asm files
in this commit.

Finished commands:
add
sub
neg
and
or