Project 7: code works for simple add (push, add)
Got the code working to write out assembly for a push and add VM
commands. I tested out the output for the SimpleAdd.vm file and it works
correctly. Most of the logic is now written for the VM translator, I
need to add the glue code for the other VM commands