Project 8: Create VM translator part 2 - handle branching commands
This commit handles translating the label, goto, and if-goto (AKA branching
commands) from VM language to assembly code. This was tested out on the
two "ProgramFlow" tests, BasicLoop and FibonacciSeries, and both worked
in the CPU emulator. This work seemed easier than part 1 of the VM
translator last week since I'm getting more fluid with the assembly code.
There's still a bunch more to this project.