Refactor tokenizer to parse all tokens up-front
I was trying to write a "peek" method for the CompilationEngine and it
was turning out to be more complicated than needed so I made this
refactor as a pre-step.

This change will now parse all the tokens in one pass rather than parsing them as needed. This makes it
easier since we don't need to track where we are in the text of the file and makes
the logic for methods used by compilation engine (advance, peek_at_next_token)
much simpler.