Finished CompilationEngine - now compiles all test files correctly
* Got expressions/terms compiling correctly which was the most difficult
part since we needed to add a look-ahead for certain types of terms.
* Added some small changes to the tokenizer such as removing tabs from
the input file and converting < > " & to use non-special characters
which don't interfere with the XML.
* The test file is bringing in full test files from SquareGame,
compiling them, and comparing each line to the XML output from the test
directory. They all pass :) So at this point I'm testing on over 2k
lines of compiled XML and am confident my compilation engine is working.