Refactor labels/goto/if-goto names to be more intuitive
* Previously I was using an incrementing integer to keep ensure label
names (for example "label L1", "goto L4") which worked but which didn't
distinguish between labels used for "if" and "while" loops and which got
confusing when the count got higher. This change makes label naming
call out the loop type and branch (for example "label IF_TRUE0", "goto
WHILE_EXP1").