Re-compile projects with update compiler
* My Jack compiler is now completely consistent with the built-in
compiler that comes in the tools directory (JackCompiler.sh) so I'm
re-compiling these project files to reflect that.

* This makes testing much easier because I can now compile the projects using both the
built-in compiler and my compiler and then diff the results. If they're
the same I know my compiler has no bugs. I'm still testing each compiled
program in the VM emulator, but diffing the compiler output is a good
sanity check.