Refactor VM output object to live in VM writer
* Previously there was a VM output object on the Compilation Object
class which was an array storing all VM commands. I was using it for
testing - to compare VM command snippets. It was a bit messy
having all the VM commands get returned from the VM writer to be
appended to that array. This refactor moves that object into the VM
writer class which makes more sense conceptionally. A further possible
refactor is to write all VM commands immediately to a file (rather
than to a var) and then compare them in the outputted file.