Refactor compile term
* I wasn't happy with the way that compile_term worked differently from
the other compile_* methods. Specifically it returned a list of elements
that made up the term and which were then converted into VM commands by
the code_write method. This was clunky and hard to reason through. I had
trouble getting it to work for the ComplexArrays test project. So in
this commit I refactored compile_term to handle the compilation itself,
rather than returning a list of elements to be converted into VM
commands, and removing the need for the code_write method altogether.
I'm really happy with this refactor and feel it makes the code stronger
and cleaner.